+++ apps/zoos/src/features/objects/db-interface.ts
