+++ apps/zoos/src/features/objects/db/schema.sql
