+++ apps/zoos/src/features/objects/types.ts
